1.차량 운전석에서만 가방 내 아이템 이동 불가능하게 변경 2. 바리케이드 보급품 추가 3. 거점 점령 점수 계산식 수정 4. 즉사하면 킬로그 안남는 버그 수정

+++ KoreanWarPlugin/Info/ObjectiveInfo.cs
@@ -27,8 +27,8 @@ namespace KoreanWarPlugin.Info
         public SpawnPreset[] objectiveSpawn;
         public SpawnPreset[] team_0_spawn;
         public SpawnPreset[] team_1_spawn;
-        public Dictionary<SteamPlayerID, ushort> team_0_effort; // 거점에서 점수를 얻은 0팀 유저들
-        public Dictionary<SteamPlayerID, ushort> team_1_effort; // 거점에서 점수를 얻은 1팀 유저들
+        public Dictionary<SteamPlayerID, float> team_0_effort; // 거점에서 점수를 얻은 0팀 유저들
+        public Dictionary<SteamPlayerID, float> team_1_effort; // 거점에서 점수를 얻은 1팀 유저들
         public ObjectiveInfo(float _point, ObjectivePreset _objectivePreset)
         {
             point = _point;
@@ -47,8 +47,8 @@ namespace KoreanWarPlugin.Info
             objectiveSpawn = _objectivePreset.objectiveSpawn;
             team_0_spawn = _objectivePreset.team_0_spawn;
             team_1_spawn = _objectivePreset.team_1_spawn;
-            team_0_effort = new Dictionary<SteamPlayerID, ushort>();
-            team_1_effort = new Dictionary<SteamPlayerID, ushort>();
+            team_0_effort = new Dictionary<SteamPlayerID, float>();
+            team_1_effort = new Dictionary<SteamPlayerID, float>();
         }
     }
 }
